backend/pyproject.toml
[tool.poetry]
name = "fast"
version = "0.1.0"
description = ""
authors = ["DeveloperInal <166642110+DeveloperInal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
loguru = "^0.7.2"
aiogram = "^3.13.1"
gunicorn = "^23.0.0"
prometheus-client = "^0.21.0"
prometheus-fastapi-instrumentator = "^7.0.0"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
black = "^25.1.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
